requirements.txt
Django~=4.2.3
djangorestframework~=3.14.0
python-dotenv~=1.0.0
black~=23.7.0
psycopg2-binary~=2.9.6
djangorestframework-simplejwt~=5.2.2
pre-commit~=3.3.3
django-phonenumber-field[phonenumberslite]~=7.1.0
gunicorn~=21.2.0
django-filter==23.2
